Streamlit/requirements.txt
pandas==2.3.1
Pillow==11.3.0
streamlit==1.48.0
os